refactor: move the hooks to an appropriate folder
