+++ src/hooks/index.ts
@@ -1,0 +1,7 @@
+import useClickOutside from './useClickOutside';
+import useScroll from './useScroll';
+import useTheme from './useTheme';
+import useToggle from './useToggle';
+import useTranslation from './useTranslation';
+
+export { useClickOutside, useScroll, useTheme, useToggle, useTranslation };
