+++ src/pages/ErrorPage.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Section from '../components/Section';
 import { pageNotFound } from '../data/pageNotFound';
+import useTranslation from '../hooks/useTranslation';
 import CustomLayout from '../layout/CustomLayout';
-import useTranslation from '../utils/useTranslation';
 
 const ErrorPage = () => {
   const { t } = useTranslation();
