--- src/utils/useNavbar.ts
@@ -1,12 +1,0 @@
-import { useState } from 'react';
-
-const useNavbar = () => {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const toggleNavbar = () => {
-    setIsOpen(!isOpen);
-  };
-  return [isOpen, toggleNavbar];
-};
-
-export default useNavbar;
